# swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Some custom description of API.",
		"version": "API TOS",
		"title": "My REST API",
		"termsOfService": "Terms of service",
		"contact": {
			"name": "John Doe",
			"url": "www.example.com",
			"email": "myeaddress@company.com"
		},
		"license": {
			"name": "License of API",
			"url": "API license URL"
		}
	},
	"host": "localhost:9191",
	"basePath": "/ProApp2",
...
	"paths": {
		"/delete": {
			"get": {
				"tags": [
					"form-submittion"
				],
				"summary": "viewDelete",
				"operationId": "viewDeleteUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"head": {
				"tags": [
					"form-submittion"
				],
				"summary": "viewDelete",
				"operationId": "viewDeleteUsingHEAD",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"post": {
				"tags": [
					"form-submittion"
				],
				"summary": "viewDelete",
				"operationId": "viewDeleteUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"put": {
				"tags": [
					"form-submittion"
				],
				"summary": "viewDelete",
				"operationId": "viewDeleteUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"delete": {
				"tags": [
					"form-submittion"
				],
				"summary": "viewDelete",
				"operationId": "viewDeleteUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"options": {
				"tags": [
					"form-submittion"
				],
				"summary": "viewDelete",
				"operationId": "viewDeleteUsingOPTIONS",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"patch": {
				"tags": [
					"form-submittion"
				],
				"summary": "viewDelete",
				"operationId": "viewDeleteUsingPATCH",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			}
		},
		"/get": {
			"get": {
				"tags": [
					"form-submittion"
				],
				"summary": "getForm",
				"operationId": "getFormUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"head": {
				"tags": [
					"form-submittion"
				],
				"summary": "getForm",
				"operationId": "getFormUsingHEAD",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"post": {
				"tags": [
					"form-submittion"
				],
				"summary": "getForm",
				"operationId": "getFormUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"put": {
				"tags": [
					"form-submittion"
				],
				"summary": "getForm",
				"operationId": "getFormUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"delete": {
				"tags": [
					"form-submittion"
				],
				"summary": "getForm",
				"operationId": "getFormUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"options": {
				"tags": [
					"form-submittion"
				],
				"summary": "getForm",
				"operationId": "getFormUsingOPTIONS",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"patch": {
				"tags": [
					"form-submittion"
				],
				"summary": "getForm",
				"operationId": "getFormUsingPATCH",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			}
		},
...
		"/registerUser": {
			"post": {
				"tags": [
					"form-submittion"
				],
				"summary": "save",
				"operationId": "saveUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "country",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "email",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "flashAttributes",
						"in": "query",
						"required": false,
						"type": "object"
					},
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "name",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "phone",
						"in": "query",
						"required": false,
						"type": "integer",
						"format": "int64"
					},
					{
						"name": "pwd",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "userId",
						"in": "query",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/result": {
			"get": {
				"tags": [
					"form-submittion"
				],
				"summary": "redirect",
				"operationId": "redirectUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"head": {
				"tags": [
					"form-submittion"
				],
				"summary": "redirect",
				"operationId": "redirectUsingHEAD",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"post": {
				"tags": [
					"form-submittion"
				],
				"summary": "redirect",
				"operationId": "redirectUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"put": {
				"tags": [
					"form-submittion"
				],
				"summary": "redirect",
				"operationId": "redirectUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"delete": {
				"tags": [
					"form-submittion"
				],
				"summary": "redirect",
				"operationId": "redirectUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"options": {
				"tags": [
					"form-submittion"
				],
				"summary": "redirect",
				"operationId": "redirectUsingOPTIONS",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"patch": {
				"tags": [
					"form-submittion"
				],
				"summary": "redirect",
				"operationId": "redirectUsingPATCH",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			}
		},
		"/update": {
			"get": {
				"tags": [
					"form-submittion"
				],
				"summary": "update",
				"operationId": "updateUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"head": {
				"tags": [
					"form-submittion"
				],
				"summary": "update",
				"operationId": "updateUsingHEAD",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"post": {
				"tags": [
					"form-submittion"
				],
				"summary": "update",
				"operationId": "updateUsingPOST_1",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"put": {
				"tags": [
					"form-submittion"
				],
				"summary": "update",
				"operationId": "updateUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"delete": {
				"tags": [
					"form-submittion"
				],
				"summary": "update",
				"operationId": "updateUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"options": {
				"tags": [
					"form-submittion"
				],
				"summary": "update",
				"operationId": "updateUsingOPTIONS",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			},
			"patch": {
				"tags": [
					"form-submittion"
				],
				"summary": "update",
				"operationId": "updateUsingPATCH",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			}
		},
		"/updateUser": {
			"post": {
				"tags": [
					"form-submittion"
				],
				"summary": "update",
				"operationId": "updateUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "country",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "email",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "map",
						"in": "query",
						"description": "map",
						"required": false,
						"type": "object"
					},
					{
						"name": "name",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "phone",
						"in": "query",
						"required": false,
						"type": "integer",
						"format": "int64"
					},
					{
						"name": "pwd",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					},
					{
						"name": "userId",
						"in": "query",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
	}
}